+++ src/util/dates.test.js
@@ -1,16 +1,30 @@
-import { fakeIntl } from './test-data';
+import { createIntl, createIntlCache } from './reactIntl';
 import {
   isDate,
   isSameDate,
-  nightsBetween,
   daysBetween,
   minutesBetween,
-  formatDate,
+  diffInTime,
+  formatDateWithProximity,
+  formatDateIntoPartials,
   parseDateFromISO8601,
   stringifyDateToISO8601,
+  monthIdString,
 } from './dates';
 
 describe('date utils', () => {
+  const cache = createIntlCache();
+  const intl = createIntl(
+    {
+      locale: 'en-US',
+      messages: {},
+    },
+    cache
+  );
+  // Tests inject now() function to intl wich returns predefined date
+  // Note: fakeIntl uses this same moment
+  intl.now = () => Date.UTC(2017, 10, 23, 12, 59);
+
   describe('isDate()', () => {
     it('should return false if parameters is string', () => {
       expect(isDate('Monday')).toBeFalsy();
@@ -34,28 +48,6 @@ describe('date utils', () => {
     });
     it('should be truthy if parameters match', () => {
       expect(isSameDate(new Date(2019, 0, 1), new Date(2019, 0, 1))).toBeTruthy();
-    });
-  });
-
-  describe('nightsBetween()', () => {
-    it('should fail if end date is before start date', () => {
-      const start = new Date(2017, 0, 2);
-      const end = new Date(2017, 0, 1);
-      expect(() => nightsBetween(start, end)).toThrow('End date cannot be before start date');
-    });
-    it('should handle equal start and end dates', () => {
-      const d = new Date(2017, 0, 1);
-      expect(nightsBetween(d, d)).toEqual(0);
-    });
-    it('should calculate night count for a single night', () => {
-      const start = new Date(2017, 0, 1);
-      const end = new Date(2017, 0, 2);
-      expect(nightsBetween(start, end)).toEqual(1);
-    });
-    it('should calculate night count', () => {
-      const start = new Date(2017, 0, 1);
-      const end = new Date(2017, 0, 3);
-      expect(nightsBetween(start, end)).toEqual(2);
     });
   });
 
@@ -103,19 +95,54 @@ describe('date utils', () => {
     });
   });
 
-  describe('formatDate()', () => {
-    /*
-      NOTE: These are not really testing the formatting properly since
-      the fakeIntl object has to be used in the tests.
-     */
+  describe('diffInTime()', () => {
+    it('should calculate 18 years between 2000-01-01 and 2018-01-01', () => {
+      const birth = new Date(Date.UTC(2000, 0, 1));
+      const now = new Date(Date.UTC(2018, 0, 1));
+      expect(diffInTime(now, birth, 'years', true)).toEqual(18);
+    });
+  });
 
+  describe('formatDateWithProximity()', () => {
     it('formats a date today', () => {
       const d = new Date(Date.UTC(2017, 10, 23, 13, 51));
-      expect(formatDate(fakeIntl, 'Today', d)).toEqual('Today, 13:51');
+      expect(formatDateWithProximity(d, intl, 'Today', { timeZone: 'Etc/UTC' })).toEqual(
+        'Today, 1:51 PM'
+      );
     });
-    it('formats a date', () => {
+    it('formats a date on same week', () => {
       const d = new Date(Date.UTC(2017, 10, 22, 13, 51));
-      expect(formatDate(fakeIntl, 'Today', d)).toEqual('2017-11-22, 13:51');
+      expect(formatDateWithProximity(d, intl, 'Today', { timeZone: 'Etc/UTC' })).toEqual(
+        'Wed 1:51 PM'
+      );
+    });
+    it('formats a date on same year', () => {
+      const d = new Date(Date.UTC(2017, 10, 2, 13, 51));
+      expect(formatDateWithProximity(d, intl, 'Today', { timeZone: 'Etc/UTC' })).toEqual(
+        'Nov 2, 1:51 PM'
+      );
+    });
+    it('formats a date on different year', () => {
+      const d = new Date(Date.UTC(2020, 10, 2, 13, 51));
+      expect(formatDateWithProximity(d, intl, 'Today', { timeZone: 'Etc/UTC' })).toEqual(
+        'Nov 2, 2020, 1:51 PM'
+      );
+    });
+    it('formats 2017-11-23 00:00 UTC as "Wed 7:00 PM" in New York tz. (I.e. not as "Today, 12:00 AM")', () => {
+      const d = new Date(Date.UTC(2017, 10, 23, 0, 0));
+      expect(formatDateWithProximity(d, intl, 'Today', { timeZone: 'America/New_York' })).toEqual(
+        'Wed 7:00 PM'
+      );
+    });
+  });
+
+  describe('formatDateIntoPartials()', () => {
+    it('formats a date into its partials', () => {
+      const d = new Date(Date.UTC(2017, 10, 23, 13, 51));
+      const partials = formatDateIntoPartials(d, intl, { timeZone: 'Etc/UTC' });
+      expect(partials.time).toEqual('1:51 PM');
+      expect(partials.date).toEqual('Nov 23');
+      expect(partials.dateAndTime).toEqual('Nov 23, 1:51 PM');
     });
   });
 
@@ -125,12 +152,64 @@ describe('date utils', () => {
       const date = new Date(2018, 10, 23);
       expect(parseDateFromISO8601(dateString)).toEqual(date);
     });
+
+    it('should read ISO 8601 date as date in Etc/UTC and return date and time in UTC formatted ISO 8601', () => {
+      expect(parseDateFromISO8601('2020-04-07', 'Etc/UTC').toISOString()).toEqual(
+        '2020-04-07T00:00:00.000Z'
+      );
+    });
+
+    it('should read ISO 8601 date as date in Europe/Helsinki and return date and time in UTC formetted ISO 8601', () => {
+      expect(parseDateFromISO8601('2020-02-07', 'Europe/Helsinki').toISOString()).toEqual(
+        '2020-02-06T22:00:00.000Z'
+      );
+    });
   });
 
   describe('stringifyDateToISO8601()', () => {
     it('should return string in YYYY-MM-DD format', () => {
       const date = new Date(2018, 10, 23);
       expect(stringifyDateToISO8601(date)).toEqual('2018-11-23');
+    });
+
+    it('should read given date in Etc/UTC and return ISO 8601 date string', () => {
+      expect(stringifyDateToISO8601(new Date('2020-04-07T00:00:00.000Z'), 'Etc/UTC')).toEqual(
+        '2020-04-07'
+      );
+    });
+
+    it('should read given date in Europe/Helsinki and return ISO 8601 date string', () => {
+      expect(
+        stringifyDateToISO8601(new Date('2020-02-06T22:00:00.000Z'), 'Europe/Helsinki')
+      ).toEqual('2020-02-07');
+    });
+
+    it('should read given date in America/New_York and return ISO 8601 date string', () => {
+      const date = new Date(Date.UTC(2020, 3, 7));
+      // UTC 2020-04-07 00:00 is previous day in New York
+      expect(stringifyDateToISO8601(date, 'America/New_York')).toEqual('2020-04-06');
+    });
+  });
+
+  describe('monthIdString()', () => {
+    it('should return string in YYYY-MM format without timeZone param', () => {
+      const date = new Date(2018, 10, 23);
+      expect(monthIdString(date)).toContain('2018-11');
+    });
+
+    it('should return string in YYYY-MM format with timeZone param', () => {
+      const date = new Date(2018, 10, 23);
+      expect(monthIdString(date, 'Etc/UTC')).toEqual('2018-11');
+    });
+
+    it('should return string in YYYY-MM format with Etc/UTC tz to equal moment(date).utc().format("YYYY-MM")', () => {
+      const date = new Date(2018, 10, 23);
+      const pad = num => {
+        return num >= 0 && num < 10 ? `0${num}` : `${num}`;
+      };
+      expect(monthIdString(date, 'Etc/UTC')).toEqual(
+        `${date.getUTCFullYear()}-${pad(date.getUTCMonth() + 1)}`
+      );
     });
   });
 });